CAMS-142: Separated http types and created test case for code coverage
