+++ backend/functions/lib/adapters/utils/http-client.ts
@@ -1,28 +1,4 @@
-import { ApiResponse, HttpResponse } from '../types/http';
-import { CamsError } from '../../common-errors/cams-error';
-
-const commonHeaders = {
-  'Content-Type': 'application/json',
-  'Last-Modified': Date.toString(),
-};
-
-export function httpSuccess(body: object = {}): ApiResponse {
-  return {
-    headers: commonHeaders,
-    statusCode: 200,
-    body,
-  };
-}
-
-export function httpError(error: CamsError): ApiResponse {
-  return {
-    headers: commonHeaders,
-    statusCode: error.status,
-    body: {
-      error: error.message,
-    },
-  };
-}
+import { HttpResponse } from '../types/http';
 
 // fetch post
 export async function httpPost(data: {
